+++ server/graphql/invoice_line/src/mutations/outbound_shipment_line/unallocated_line/update.rs
@@ -18,7 +18,7 @@ use service::{
 #[graphql(name = "UpdateOutboundShipmentUnallocatedLineInput")]
 pub struct UpdateInput {
     pub id: String,
-    pub quantity: u32,
+    pub quantity: f64,
 }
 
 #[derive(Interface)]
@@ -90,7 +90,7 @@ fn map_error(error: ServiceError) -> Result<UpdateErrorInterface> {
             return Ok(UpdateErrorInterface::RecordNotFound(RecordNotFound {}))
         }
         ServiceError::InvoiceDoesNotExist => {
-            //TODO: Change all to std error or update check_line_exists_option
+            //TODO: Change all to std error or update check_line_exists
             //https://github.com/openmsupply/open-msupply/pull/366#discussion_r930574975
             return Ok(UpdateErrorInterface::ForeignKeyError(ForeignKeyError(
                 ForeignKey::InvoiceId,
